make eventslist component
