+++ frontend/src/Pages/EventsPage/EventsPage.tsx
@@ -1,20 +1,16 @@
 import { useEffect, useState } from 'react';
-import { Link } from 'react-router-dom';
 import styles from './EventsPage.module.scss';
 import { ROUTES } from 'routes';
 import { Event } from 'types';
+import { EventsList } from './components/EventsList';
 
-/** check if dates are equal */
-function compareDates(event_date: Date, array_date: string) {
-  return event_date.toISOString().includes(array_date);
-}
 /** get date on format hh.mm */
 export const getTimeStr = (date: Date) => {
   return (date.getHours() + '.' + date.getMinutes()).toString();
 };
 
 export function EventsPage() {
-  const [events, setEvents] = useState([]);
+  const [events, setEvents] = useState<Event[]>([]);
   const [dates, setDates] = useState<string[]>([]);
   useEffect(() => {
     fetch(ROUTES.backend.events)
@@ -33,41 +29,8 @@ export function EventsPage() {
       });
   }, []);
   return (
-    <div className={styles.container}>
-      {dates.map((date_str: string, index: number) => {
-        const date: Date = new Date(date_str);
-        const month: string = date.toLocaleDateString('no', { month: 'long' });
-        const weekday: string = date.toLocaleDateString('no', { weekday: 'long' });
-        const monthday: number = date.getDate();
-
-        return (
-          <div key={index} className={styles.dates_container}>
-            <h2 className={styles.date_header}>{weekday + ' ' + monthday + '.' + month}</h2>
-            {events.map((event: Event, index: number) => {
-              if (compareDates(event.start_dt, date_str)) {
-                return (
-                  <div className={styles.events_container} key={index}>
-                    {' '}
-                    <div className={styles.event_row}>
-                      <div className={styles.column_title}>
-                        <Link to={'/events/' + event.id}>
-                          <p> {event.title_no}</p>{' '}
-                        </Link>
-                      </div>
-                      <div className={styles.column_area_time}>
-                        <div className={styles.time_wrapper}>
-                          <p>{getTimeStr(event?.start_dt) + ' - ' + getTimeStr(event?.end_dt)}</p>
-                        </div>
-                        <p> {event?.location}</p>
-                      </div>
-                    </div>
-                  </div>
-                );
-              }
-            })}
-          </div>
-        );
-      })}
+    <div className={styles.root_container}>
+      <EventsList event_list={events} date_list={dates} />
     </div>
   );
 }
