+++ frontend/src/Pages/EventsPage/components/EventsList.module.scss
@@ -1,0 +1,69 @@
+body {
+  font-family: 'futura-pt', 'Lato', sans-serif;
+}
+
+.date_header {
+  background-color: #a03033;
+  width: 100%;
+  text-align: center;
+  color: white;
+  font-weight: 500;
+  font-size: 1.2em;
+  line-height: 1.2;
+  text-transform: uppercase;
+}
+
+.container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  max-width: 90%;
+  width: 90%;
+  background-color: transparent;
+  margin: 16px;
+  height: 100%;
+}
+
+.dates_container {
+  width: 100%;
+  justify-content: space-between;
+  display: flex;
+  flex-direction: column;
+  box-shadow: 2px 2px 15px 4px rgb(0 0 0 / 10%);
+  border-radius: 0.5em;
+  overflow: hidden;
+  align-items: center;
+  margin-bottom: 16px;
+}
+
+.events_container {
+  display: flex;
+  flex-direction: row;
+  overflow: hidden;
+  align-items: center;
+  width: 100%;
+  font-size: 0.8em;
+  padding: 0.2em;
+  padding-left: 1em;
+  padding-right: 1em;
+}
+
+.column_area_time {
+  flex: 1;
+}
+
+.column_title {
+  flex: 1;
+  flex-direction: row;
+  justify-content: flex-end;
+}
+
+.time_wrapper {
+}
+
+.event_row {
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+}
