+++ frontend/src/Pages/EventsPage/components/index.ts
@@ -1,0 +1,1 @@
+export { EventsList } from './EventsList';
